FEATURE: No Monthly Balances  (#14)
* feat: flow without mbs

* feat: design time db context factory

* feat: spreadsheet service for no mbs flow

* refactor: new name for spreadsheet service

* fix: deprecated package removed

* style: code cleanup

+++ MyFinance-Backend/MyFinance.Application/UseCases/Transfers/Commands/UpdateTransfer/UpdateTransferCommand.cs
@@ -8,7 +8,7 @@ namespace MyFinance.Application.UseCases.Transfers.Commands.UpdateTransfer;
 public sealed record UpdateTransferCommand(
     Guid Id,
     Guid AccountTagId,
-    double Value,
+    decimal Value,
     string RelatedTo,
     string Description,
     DateTime SettlementDate,
